refactor: order table 객체 대신 id 참조하도록 변경
