+++ src/main/java/kitchenpos/domain/repository/OrderTableRepository.java
@@ -16,4 +16,6 @@ public interface OrderTableRepository {
     List<OrderTable> findAllByIdsIn(List<Long> ids);
 
     List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+
+    boolean existsById(long id);
 }
