+++ src/main/java/kitchenpos/dto/request/CreateProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto.request;
 
 import kitchenpos.domain.Product;
