+++ src/main/java/kitchenpos/infrastructure/persistence/OrderTableRepositoryImpl.java
@@ -40,4 +40,9 @@ public class OrderTableRepositoryImpl implements OrderTableRepository {
     public List<OrderTable> findAllByTableGroupId(final Long tableGroupId) {
         return jpaOrderTableRepository.findAllByTableGroupId(tableGroupId);
     }
+
+    @Override
+    public boolean existsById(final long id) {
+        return jpaOrderTableRepository.existsById(id);
+    }
 }
