--- src/main/java/kitchenpos/ui/dto/CreateMenuGroupRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.dto;
-
-public class CreateMenuGroupRequest {
-
-    private String name;
-
-    public CreateMenuGroupRequest() {
-    }
-
-    public CreateMenuGroupRequest(final String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
