--- src/main/java/kitchenpos/ui/dto/CreateMenuRequest.java
@@ -1,43 +1,0 @@
-package kitchenpos.ui.dto;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public class CreateMenuRequest {
-
-    private Long id;
-
-    private Long menuGroupId;
-
-    private List<MenuProductDto> menuProducts;
-
-    private String name;
-
-    private BigDecimal price;
-
-    public CreateMenuRequest() {
-    }
-
-    public CreateMenuRequest(final Long menuGroupId, final List<MenuProductDto> menuProducts, final String name, final BigDecimal price) {
-        this.menuGroupId = menuGroupId;
-        this.menuProducts = menuProducts;
-        this.name = name;
-        this.price = price;
-    }
-
-    public Long getMenuGroupId() {
-        return menuGroupId;
-    }
-
-    public List<MenuProductDto> getMenuProducts() {
-        return menuProducts;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-}
