--- src/main/java/kitchenpos/ui/dto/CreateOrderRequest.java
@@ -1,25 +1,0 @@
-package kitchenpos.ui.dto;
-
-import java.util.List;
-
-public class CreateOrderRequest {
-
-    private Long orderTableId;
-    private List<OrderLineItemDto> orderLineItems;
-
-    public CreateOrderRequest() {
-    }
-
-    public CreateOrderRequest(final Long orderTableId, final List<OrderLineItemDto> orderLineItems) {
-        this.orderTableId = orderTableId;
-        this.orderLineItems = orderLineItems;
-    }
-
-    public Long getOrderTableId() {
-        return orderTableId;
-    }
-
-    public List<OrderLineItemDto> getOrderLineItems() {
-        return orderLineItems;
-    }
-}
