--- src/main/java/kitchenpos/ui/dto/CreateTableGroupRequest.java
@@ -1,19 +1,0 @@
-package kitchenpos.ui.dto;
-
-import java.util.List;
-
-public class CreateTableGroupRequest {
-
-    private List<OrderTableDto> orderTables;
-
-    public CreateTableGroupRequest() {
-    }
-
-    public CreateTableGroupRequest(final List<OrderTableDto> createOrderTableRequests) {
-        this.orderTables = createOrderTableRequests;
-    }
-
-    public List<OrderTableDto> getOrderTables() {
-        return orderTables;
-    }
-}
