--- src/main/java/kitchenpos/ui/dto/OrderTableDto.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.dto;
-
-public class OrderTableDto {
-
-    private long id;
-
-    public OrderTableDto() {
-    }
-
-    public OrderTableDto(final long id) {
-        this.id = id;
-    }
-
-    public long getId() {
-        return id;
-    }
-}
