--- src/main/java/kitchenpos/ui/dto/PutOrderStatusRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.dto;
-
-public class PutOrderStatusRequest {
-
-    private String orderStatus;
-
-    public PutOrderStatusRequest() {
-    }
-
-    public PutOrderStatusRequest(final String orderStatus) {
-        this.orderStatus = orderStatus;
-    }
-
-    public String getOrderStatus() {
-        return orderStatus;
-    }
-}
