--- src/main/java/kitchenpos/ui/dto/PutOrderTableEmptyRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.dto;
-
-public class PutOrderTableEmptyRequest {
-
-    private boolean empty;
-
-    public PutOrderTableEmptyRequest() {
-    }
-
-    public PutOrderTableEmptyRequest(final boolean empty) {
-        this.empty = empty;
-    }
-
-    public boolean getEmpty() {
-        return empty;
-    }
-}
