--- src/main/java/kitchenpos/ui/dto/PutOrderTableGuestsNumberRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.dto;
-
-public class PutOrderTableGuestsNumberRequest {
-
-    private int numberOfGuests;
-
-    public PutOrderTableGuestsNumberRequest() {
-    }
-
-    public PutOrderTableGuestsNumberRequest(final int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
