--- src/main/java/kitchenpos/ui/response/MenuGroupResponse.java
@@ -1,23 +1,0 @@
-package kitchenpos.ui.response;
-
-public class MenuGroupResponse {
-
-    private Long id;
-    private String name;
-
-    public MenuGroupResponse() {
-    }
-
-    public MenuGroupResponse(final Long id, final String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
