--- src/main/java/kitchenpos/ui/response/OrderLineItemDtoInOrderResponse.java
@@ -1,17 +1,0 @@
-package kitchenpos.ui.response;
-
-public class OrderLineItemDtoInOrderResponse {
-
-    private Long id;
-
-    public OrderLineItemDtoInOrderResponse() {
-    }
-
-    public OrderLineItemDtoInOrderResponse(final Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
