--- src/main/java/kitchenpos/ui/response/OrderTableResponse.java
@@ -1,37 +1,0 @@
-package kitchenpos.ui.response;
-
-import kitchenpos.domain.TableGroup;
-
-import java.util.Objects;
-
-public class OrderTableResponse {
-
-    private Long id;
-    private Long tableGroupId;
-    private int numberOfGuests;
-    private boolean empty;
-
-    public OrderTableResponse() {
-    }
-
-    public OrderTableResponse(final Long id, final TableGroup tableGroup, final int numberOfGuests, final boolean empty) {
-        this.id = id;
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-        if (Objects.nonNull(tableGroup)) {
-            this.tableGroupId = tableGroup.getId();
-        }
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getTableGroupId() {
-        return tableGroupId;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
